TEST/TEST/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
Pillow==10.4.0
numpy==1.26.4
easyocr==1.7.1
torch==2.3.1
PyMuPDF==1.24.7
opencv-python-headless==4.10.0.84 